lotteon/수정

+++ lotteon/src/main/java/kr/co/lotteon/entity/Review.java
@@ -28,7 +28,8 @@ public class Review {
     private Double rating;
 
     private String Users_uid;
-    private int rating;
+
+    private Double rating;
 
     private String comment;
